Server: Move MCC routes into own module
